optimize colorizers
- synch common methods (overrides, config, categories selection) and use same methods from the modules database and colorize
- fix issues with double parameters in Filters by Value
