+++ lib/pychilizer/colorize.py
@@ -142,3 +142,60 @@ def get_colours(n):
     for x in range(10):
         random.shuffle(revit_colours)
     return revit_colours
+
+
+override_options = ["Projection Line Colour", "Projection Surface Colour", "Cut Line Colour", "Cut Pattern Colour"]
+default_options = ["Projection Surface Colour", "Cut Pattern Colour"]
+
+class ChosenItem(forms.TemplateListItem):
+    """Wrapper class for chosen item"""
+    pass
+
+def get_config(custom_config):
+    prev_choice = custom_config.get_option("overrides", [])
+    if not prev_choice:
+        save_config([x for x in default_options], custom_config)
+        prev_choice = custom_config.get_option("overrides", [])
+    return prev_choice
+
+def save_config(chosen, config):
+    """Save given list of overrides"""
+    config.overrides = chosen
+    script.save_config()
+
+def load_configs(config):
+    """Load list of frequently selected items from configs or defaults"""
+    ovrds = config.get_option("overrides", [])
+    ovrd_items = [x for x in (ovrds or default_options)]
+    if not ovrds:
+        ovrd_items = [x for x in default_options]
+    return filter(None, ovrd_items)
+
+
+def config_overrides(config):
+    """Ask for users choice of overrides"""
+    prev_ovrds = load_configs(config)
+    opts = [ChosenItem(x, checked=x in prev_ovrds) for x in override_options]
+    overrides = forms.SelectFromList.show(
+        opts,
+        title="Choose Overrides Styles",
+        button_name="Remember",
+        multiselect=True
+    )
+    if overrides:
+        save_config([x for x in overrides if x], config)
+
+def set_colour_overrides_by_option (overrides_option, colour, doc):
+    override = DB.OverrideGraphicSettings()
+    solid_fill_pat_id = database.get_solid_fill_pat(doc).Id
+    if "Projection Line Colour" in overrides_option:
+        override.SetProjectionLineColor(colour)
+    if "Cut Line Colour" in overrides_option:
+        override.SetCutLineColor(colour)
+    if "Projection Surface Colour" in overrides_option:
+        override.SetSurfaceForegroundPatternColor(colour)
+        override.SetSurfaceForegroundPatternId(solid_fill_pat_id)
+    if "Cut Pattern Colour" in overrides_option:
+        override.SetCutForegroundPatternColor(colour)
+        override.SetCutForegroundPatternId(solid_fill_pat_id)
+    return override
